removed security classes from automatic
