+++ composer.json
@@ -47,7 +47,10 @@
         "branch-alias": {
             "dev-master": "1.0-dev"
         },
-        "class": "Narrowspark\\Automatic\\Automatic"
+        "class": [
+            "Narrowspark\\Automatic\\Security\\SecurityPlugin",
+            "Narrowspark\\Automatic\\Automatic"
+        ]
     },
     "autoload": {
         "psr-4": {
