+++ src/Automatic/Container.php
@@ -55,6 +55,9 @@ final class Container implements ContainerContract
             Composer::class => static function () use ($composer) {
                 return $composer;
             },
+            Config::class => static function (Container $container) {
+                return $container->get(Composer::class)->getConfig();
+            },
             IOInterface::class => static function () use ($io) {
                 return $io;
             },
@@ -77,9 +80,6 @@ final class Container implements ContainerContract
             },
             Lock::class => static function () {
                 return new Lock(Util::getAutomaticLockFile());
-            },
-            Config::class => static function (Container $container) {
-                return $container->get(Composer::class)->getConfig();
             },
             ClassFinder::class => static function (Container $container) {
                 return new ClassFinder($container->get('vendor-dir'));
