+++ src/Security/Downloader/AbstractDownloader.php
@@ -2,7 +2,7 @@
 declare(strict_types=1);
 namespace Narrowspark\Automatic\Security\Downloader;
 
-use Narrowspark\Automatic\Security\AuditPlugin;
+use Narrowspark\Automatic\Security\SecurityPlugin;
 use Narrowspark\Automatic\Security\Contract\Downloader as DownloaderContract;
 use Narrowspark\Automatic\Security\Contract\Exception\RuntimeException;
 
@@ -30,7 +30,7 @@ abstract class AbstractDownloader implements DownloaderContract
     {
         return \sprintf(
             'Narrowspark-Automatic/%s (%s; %s; %s%s)',
-            AuditPlugin::VERSION,
+            SecurityPlugin::VERSION,
             \function_exists('php_uname') ? \php_uname('s') : 'Unknown',
             \function_exists('php_uname') ? \php_uname('r') : 'Unknown',
             'PHP ' . \PHP_MAJOR_VERSION . '.' . \PHP_MINOR_VERSION . '.' . \PHP_RELEASE_VERSION,
